Move dotnet-fpe repo into dotnet repo

See merge request ubiqsecurity/ubiq-dotnet!22

Former-commit-id: 2cd379b11c21f2fdc7eed4dcfd280b894e221ce4

+++ src/UbiqSecurity/Billing/BillingEventsManager.cs
@@ -25,7 +25,7 @@ namespace UbiqSecurity.Billing
 			_ubiqWebService = ubiqWebService;
 		}
 
-		public int EventCount => _billingEvents.Count;
+		public long EventCount => _billingEvents.Count;
 
 		public async Task AddBillingEventAsync(string apiKey, string datasetName, string datasetGroupName, BillingAction billingAction, DatasetType datasetType, int keyNumber, long count)
 		{
